frontend/src/components/Chatbot/carbon-chatbot/requirements.txt
flask
flask-cors
llama-index
llama-index-embeddings-huggingface
llama-index-llms-ollama
llama-index-vector-stores-faiss
faiss-cpu
PyPDF2
pdfplumber
googletrans==4.0.0rc1
matplotlib
pandas
python-dotenv
ollama
reportlab
seaborn
numpy
scikit-learn
plotly
kaleido

# Enhanced dependencies for 10x performance
groq
redis
PyMuPDF
pdf2image
pytesseract
sentence-transformers
requests
beautifulsoup4
libretranslate
transformers
torch
accelerate
weasyprint
Pillow
asyncio
concurrent.futures
spellchecker
textstat
nltk
